+++ flake.nix
@@ -2,6 +2,10 @@
   inputs = {
     nixpkgs.url = "github:NixOS/nixpkgs/bc16855ba53f3cb6851903a393e7073d1b5911e7";
     flake-utils.url = "github:numtide/flake-utils";
+    adf = {
+        url = "github:yaitskov/add-dependent-file";
+        flake = false;
+      };
     uphack = {
       url = "github:yaitskov/upload-doc-to-hackage";
       flake = false;
@@ -22,7 +26,9 @@
         };
         packageName = "css-class-bindings";
         pkgs = nixpkgs.legacyPackages.${system};
-        haskellPackages = pkgs.haskell.packages.${ghcName};
+        haskellPackages = pkgs.haskell.packages.${ghcName}.extend(final: prev: {
+          add-dependent-file = final.callCabal2nix "add-dependent-file" inputs.adf { };
+        });
       in {
         packages.default =
             haskellPackages.callCabal2nix packageName (sourceFilter ./.) {};