+++ src/CssClassBindings/IncludeCss.hs
@@ -1,21 +1,18 @@
 module CssClassBindings.IncludeCss (includeCss) where
 
+import AddDependentFile ( (</>), addDependentFile, getPackageRoot )
 import CssClassBindings.Escape ( escapeValIden )
 import CssClassBindings.Qq (cssToDecs)
 import Data.Text.IO.Utf8 qualified as U
-import Language.Haskell.TH.Syntax
-    ( mkName, Q, Dec, runIO, getPackageRoot )
+import Language.Haskell.TH.Syntax ( mkName, Q, Dec, runIO )
 import Prelude
-import System.Directory (makeAbsolute)
-import System.FilePath (takeBaseName, (</>))
-
-liftIO1 :: (a -> IO b) -> a -> Q b
-liftIO1 f x = runIO (f x)
+import System.FilePath (takeBaseName)
 
 -- | like css quasi quoter but
 -- css input is exported via constant equal to base file name
 -- instead of cssAsLiteralText.
 includeCss :: FilePath -> Q [Dec]
 includeCss p = do
-  ap <- getPackageRoot >>= liftIO1 makeAbsolute . (</> p)
+  ap <- (</> p) <$> getPackageRoot
+  addDependentFile ap
   cssToDecs (mkName (escapeValIden $ takeBaseName p)) <$> runIO (U.readFile ap)