+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 0.0.3 2026-03-30
   * generate unit types for CSS identifiers
+  * includeCss marks files with addDependentFile
 
 ## Version 0.0.2 2026-03-29
   * includeCss